refactor: update import paths and restructure project organization
- Changed import paths from `insurance_product_data_spain` to `src` for better modularity and clarity.
- Introduced new core and schema modules to enhance project structure and maintainability.
- Added versioning and initialization files to support the new package organization.
