+++ scripts/process_pdfs.py
@@ -13,7 +13,7 @@ import argparse
 from datetime import datetime, timezone
 from pathlib import Path
 
-from insurance_product_data_spain.core.logging import logger
+from src.core.logging import logger
 from scripts.config import INSURANCE_COMPANIES_JSON, INSURANCE_COMPANIES_STORAGE_PATH
 from scripts.utils.pdf_processing import PDFConversionResult, convert_pdf_to_markdown
 from scripts.utils.storage import ensure_directory, load_json, save_json
