+++ scripts/sync_folders.py
@@ -8,7 +8,7 @@ Run with: uv run python -m scripts.sync_folders
 
 from pathlib import Path
 
-from insurance_product_data_spain.core.logging import logger
+from src.core.logging import logger
 from scripts.config import INSURANCE_COMPANIES_JSON, INSURANCE_COMPANIES_STORAGE_PATH
 from scripts.utils.storage import ensure_directory, load_json
 from scripts.utils.text_transformations import text_to_snake_case
